pyproject.toml
[project]
name = "titan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioimaplib>=2.0.1",
    "aiosmtplib>=5.0.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "celery>=5.6.2",
    "fast-bitrix24>=1.8.11",
    "fastapi>=0.128.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-docx>=1.2.0",
    "redis>=7.1.0",
    "reportlab>=4.4.9",
    "respx>=0.22.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
